Change return type of penguin_bravo to short
This conforms with the second exercise and will be used to show
CI pipeline failing. I also changed the interfaces.h to prevent
the CI from failing.

+++ penguin/bravo.c
@@ -5,7 +5,7 @@
  * to something other than zero.
  */
 
-int penguin_bravo(void) {
+short penguin_bravo(void) {
    return 0;
 }
